use BorderPane as layout for wizard
